requirements.txt
django
django-debug-toolbar
mysqlclient
djangorestframework
drf-nested-routers
pillow
djoser
djangorestframework-simplejwt